split out EpochPollRunnable
